requirements.txt
Flask==2.1.0
openai==0.27.0
requests==2.28.1
mysql-connector-python==8.0.33
Werkzeug==2.2.2
